pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paradox-machine"
version = "0.1.0"
description = "A structured paradox detection toolkit built on multi-phase LLM reasoning."
readme = "README.md"
requires-python = ">=3.10"
keywords = ["llm", "paradox", "reasoning", "analysis", "prompt-engineering"]
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
]

[project.scripts]
paradox-machine = "run:main"
paradox-demo = "src.demo:main"

[tool.setuptools]
py-modules = ["run"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
namespaces = true